Move get-current-user and fix imports
