+++ app/account/page.tsx
@@ -2,7 +2,7 @@ import { InfoIcon } from 'lucide-react'
 import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
 
-import { getCurrentUser } from '@/actions/get-current-user'
+import { getCurrentUser } from '@/app/_actions/get-current-user'
 import Field from '@/components/form/field'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
