+++ app/suppliers/[handle]/page.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link'
 import { redirect } from 'next/navigation'
 import { ErrorBoundary } from 'react-error-boundary'
 
-import { getCurrentUser } from '@/actions/get-current-user'
+import { getCurrentUser } from '@/app/_actions/get-current-user'
 import { noTiles } from '@/components/tiles/tile-list'
 import { Badge } from '@/components/ui/badge'
 import { Button } from '@/components/ui/button'
